# doc/api.yml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Exercise 5.4
  description: "Versió OpenAPI del la solució publicada a <https://atenea.upc.edu/mod/resource/view.php?id=2158292>. Suposem que totes les crides necessiten autentificació via una api_key que indentifica un usuari determinat. Només es poc accedir als missatges de l'usuari autentificat."
servers:
  - url: 'http://localhost:8080/RevPollution/services'
paths:
...
  /logros/info:
    get:
      tags:
      - logros
      summary: Obtener información de un logro.
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: tier
        in: query
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '/home/erik/Escritorio/UNI/3/Q6/PES/BackEnd/doc/components/schemas/logro.yml'
        404:
          description: 'Error: Not Found'
        401:
          description: 'Error: Unauthorized'
        403:
          description: 'Error: Forbidden'
      security:
      - ApiKeyAuth: []
...
  /stations/info:
    get:
      tags:
      - estaciones
      summary: Obtener la informaciñón de una estación.
      parameters:
      - name: idStation
        in: query
        required: true
        schema:
          type: number
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '/home/erik/Escritorio/UNI/3/Q6/PES/BackEnd/doc/components/schemas/station.yml'
        404:
          description: 'Error: Not Found'
        401:
          description: 'Error: Unauthorized'
        403:
          description: 'Error: Forbidden'
      security:
      - ApiKeyAuth: []
...
  /stationsAdmin/info:
    get:
      tags:
      - estaciones
      summary: Obtener la informaciñón de una estación.
      parameters:
      - name: idStation
        in: query
        required: true
        schema:
          type: number
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '/home/erik/Escritorio/UNI/3/Q6/PES/BackEnd/doc/components/schemas/station.yml'
        404:
          description: 'Error: Not Found'
        401:
          description: 'Error: Unauthorized'
        403:
          description: 'Error: Forbidden'
      security:
      - ApiKeyAuth: []
...
  /userinfo:
    get:
      tags:
      - users
      summary: Obtener la informaciñón de una estación.
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '/home/erik/Escritorio/UNI/3/Q6/PES/BackEnd/doc/components/schemas/user.yml'
        404:
          description: 'Error: Not Found'
        401:
          description: 'Error: Unauthorized'
        403:
          description: 'Error: Forbidden'
      security:
      - ApiKeyAuth: []
...